+++ viewer/web/views.py
@@ -8,6 +8,8 @@
     :license: BSD, see LICENSE for more details
 """
 
+import re
+
 from flask import render_template
 from flask import g
 
@@ -28,11 +30,28 @@ def setup_git_repo():
     g.repo = git.Repo(app.config['GIT_REPO_PATH'])
 
 
+def is_ignored(branch_name):
+    """Is a branch with the given name ignored?"""
+    IGNORED_BRANCHES = app.config['GIT_BRANCHES_TO_IGNORE']
+
+    # First, check for a direct match by name.
+    if branch_name in IGNORED_BRANCHES:
+        return True
+
+    # Check for a regular-expression match (the branch names may be regular
+    # expressions).
+    for ignored_branch_re in IGNORED_BRANCHES:
+        if re.fullmatch(ignored_branch_re, branch_name) is not None:
+            return True
+
+    return False
+
+
 @app.route('/')
 def index():
     all_branches = g.repo.get_branches_on_remote(app.config['GIT_REMOTE'])
     ignored_branches = [branch for branch in all_branches
-                        if branch.name in app.config['GIT_BRANCHES_TO_IGNORE']]
+                        if is_ignored(branch.name)]
     shown_branches = [branch for branch in all_branches
                       if branch not in ignored_branches]
     git.sort_branches(shown_branches, app.config['SORT_BRANCHES_BY'])