+++ CHANGELOG.md
@@ -4,7 +4,8 @@ Changelog
 dev
 ---
 
-* -
+* Allow ignoring branches by regular expressions (`GIT_BRANCHES_TO_IGNORE` can
+  contain regular expressions).
 
 0.1 (2015-03-17)
 ----------------